server/requirements.txt
# Flask + CORS
Flask==3.1.1
flask-cors==6.0.1

# Gunicorn for production server (optional but good practice)
gunicorn

# Environment variable loading
python-dotenv==1.1.1

# Google Gemini / Generative AI
google-generativeai==0.8.5